requirements.txt
# --- Core API ---
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1

# --- Database & Graph ---
psycopg[binary]==3.2.1       # Postgres
neo4j==5.21.0                # Neo4j Python driver

# --- Messaging ---
nats-py==2.7.2               # NATS client

# --- Data / Utils ---
python-dotenv==1.0.1         # load .env files
requests==2.32.3             # basic HTTP requests
httpx==0.27.0                # async HTTP client
PyYAML==6.0.1                # YAML parsing for manifest generator

# --- Testing & Dev ---
pytest==8.3.2
pytest-asyncio==0.23.7
ruff==0.5.5
mypy==1.10.0
types-requests==2.32.0.20240712   # type stubs
